cmd/tailscale/cli: error on sandboxed KUBECONFIG
The sandboxed (App Store) variant relies on a specific temporary
entitlement to access the ~/.kube/config file.
The entitlement is only granted to specific files, and so is not
applicable to paths supplied by the user at runtime.
While there may be other ways to achieve this access to arbitrary
kubeconfig files, it's out of scope for now.

Updates #11604

Signed-off-by: Chloé Vulquin <code@toast.bunkerlabs.net>