cmd/tailscale/cli: process $KUBECONFIG as a list
Change `tailscale config kubeconfig` to work like the rest of the
ecosystem by making it write to the first extant file in the list, else
the final entry in the list.

The logic for file detection is inlined based on what `kind` does.
Technically it's a race condition, since the file could be removed/added
in between the processing steps, but the fallout shouldn't be too bad.

Using a `var` that is `=`-assigned allows for simpler in-loop logic,
since both cases are handled the same way.

Updates #11645

Signed-off-by: Chloé Vulquin <code@toast.bunkerlabs.net>